Switch from floats to doubles due to numerica issues in probability calculations in GMM

+++ grabcut/include/quantization/quantization_model.hpp
@@ -4,7 +4,7 @@
 #include <gmm/gaussian_mixture_model.hpp>
 
 struct QuantizationModel {
-    std::array<gmm::GaussianMixtureModel<float, 3>, 2> gmm;
+    std::array<gmm::GaussianMixtureModel<double, 3>, 2> gmm;
     std::array<int, 2> strongest_k;
     std::vector<std::uint8_t> component_map;
 };
